+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/android/models/AndroidModelsCommand.kt
@@ -11,7 +11,7 @@ import picocli.CommandLine
     optionListHeading = "%n@|bold,underline Options:|@%n",
     header = ["Information about available models"],
     description = ["Information about available models. For example prints list of available models to test against"],
-    subcommands = [AndroidModelsListCommand::class],
+    subcommands = [AndroidModelsListCommand::class, AndroidModelDescribeCommand::class],
     usageHelpAutoWidth = true
 )
 class AndroidModelsCommand : Runnable {